Split license ITs to speed up test execution
